python/requirements.txt
requests
beautifulsoup4
pdfminer.six
psycopg2-binary